requirements.txt
flask>=3.0.0
pinecone>=5.1.0
flasgger>=0.9.7.1
python-dotenv>=1.0.0
requests